requirements.txt
tensorflow-macos==2.15.0
keras==2.15.0
opencv-python
numpy
pillow
streamlit
pandas
matplotlib
scikit-learn
